Move private functions to private file, public to public
